data_pipeline/training/requirements.txt
# QLoRA Training Requirements
# Optimized for local GPU training (2x RTX 2080 Ti, 11GB each)

# Core ML frameworks
torch>=2.1.0
transformers>=4.36.0
accelerate>=0.25.0
datasets>=2.15.0

# PEFT (Parameter-Efficient Fine-Tuning)
peft>=0.7.0

# Quantization
bitsandbytes>=0.41.3

# Training utilities
trl>=0.7.4
sentencepiece>=0.1.99
protobuf>=4.25.0

# Evaluation
evaluate>=0.4.1
sacrebleu>=2.4.0
rouge-score>=0.1.2

# Data processing
pandas>=2.1.0
numpy>=1.24.0
pyarrow>=14.0.0

# Database
psycopg2-binary>=2.9.9

# Monitoring & Logging
tensorboard>=2.15.0
wandb>=0.16.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Model conversion (for Ollama)
# gguf>=0.1.0  # Install separately if needed

# Optional: For code execution testing
# jupyter>=1.0.0
# ipykernel>=6.27.0
